Tidied mouse code
Moved code from gl_vidrpi.c to in_rpi.c plus some tidying of
platform.pi.c to use a button array
Mouse delta is now accumulated over several events increasing
sensitivity (previous code was just wrong)
